# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.User:
    properties:
      age:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.UserResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A test RESTful API server
  title: Test API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: 获取所有用户列表
      parameters:
      - default: 1
        description: 页码，默认1
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量，默认10
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
      summary: 获取所有用户
      tags:
      - users
    post:
      consumes:
      - application/json
      description: 创建新用户
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
      summary: 创建用户
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
      summary: 删除用户
      tags:
      - users
    get:
      consumes:
      - application/json
      description: 根据ID获取用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
      summary: 获取单个用户
      tags:
      - users
    put:
      consumes:
      - application/json
      description: 更新用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
      summary: 更新用户
      tags:
      - users
swagger: "2.0"